requirements.txt
flask_ngrok
flask
better-profanity
pyspellchecker
textblob
badwords
pyenchant
#profanity_filter
datetime
